refactor: EncodingUtil에서 UrlEncoder로 이름 변경 및 파일 이동
