+++ src/main/java/com/url/shortener/service/encoder/UrlEncoder.java
@@ -1,0 +1,6 @@
+package com.url.shortener.service.encoder;
+
+public interface UrlEncoder {
+
+    String encodeFromNumber(long number);
+}
