--- src/main/java/com/url/shortener/utils/encoding/EncodingUtil.java
@@ -1,6 +1,0 @@
-package com.url.shortener.utils.encoding;
-
-public interface EncodingUtil {
-
-    String encodeFromNumber(long number);
-}
