+++ src/test/java/com/url/shortener/service/UrlEncoderTest.java
@@ -1,4 +1,4 @@
-package com.url.shortener.utils;
+package com.url.shortener.service;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,20 +8,20 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.url.shortener.utils.encoding.EncodingUtil;
+import com.url.shortener.service.encoder.UrlEncoder;
 
 @SpringBootTest
-class EncodingUtilTest {
+class UrlEncoderTest {
 
     @Autowired
-    private EncodingUtil encodingUtil;
+    private UrlEncoder urlEncoder;
 
     @ParameterizedTest
     @DisplayName("숫자가 인코딩 알고리즘을 거쳐 정상적으로 임의의 문자열로 변환된다.")
     @CsvSource(value = {"1,39C4", "123456789,ZtCQ8"})
-    void 인코딩_정상_변환_테스트(Long number, String expectedEncodedKey) {
+    void 인코딩_정상_변환_테스트(long number, String expectedEncodedKey) {
         //when
-        String actualEncodedKey = encodingUtil.encodeFromNumber(number);
+        String actualEncodedKey = urlEncoder.encodeFromNumber(number);
 
         //then
         assertThat(actualEncodedKey).isEqualTo(expectedEncodedKey);
